BackEndFlask/requirements.txt
# ==Core libs== #

# Flask framework and extentions
flask == 3.1.2
Werkzeug == 3.1.4
flask-sqlalchemy == 3.1.1
Flask-Migrate == 4.1.0
flask-marshmallow == 1.3.0
flask-jwt-extended == 4.7.1
flask_cors == 6.0.1
Flask-Limiter == 4.0.0

# Tools for serialization and validation
marshmallow >= 4.0.1
marshmallow-sqlalchemy >= 1.4.2

# SendGrid API for emails
sendgrid == 6.12.5

# Database and ORM tools
sqlalchemy == 2.0.43
pymysql == 1.1.2
redis == 6.4.0

# Security
pyjwt == 2.10.1
cryptography == 46.0.2

# Data processing
pandas == 2.3.3
openpyxl == 3.1.5
python-dotenv == 1.1.1

# Server related tools
gunicorn == 23.0.0
gevent == 25.9.1

# Testing tools
pytest == 8.4.2
postman_api == 3.0.0

# ==Transitive Dependencies== #
alembic == 1.16.5
blinker == 1.9.0
cachetools == 6.2.0
certifi == 2025.8.3
charset-normalizer == 3.4.3
click == 8.3.0
cffi == 2.0.0
Deprecated == 1.2.18
et_xmlfile == 2.0.0
google-api-core == 2.25.2
googleapis-common-protos == 1.70.0
greenlet == 3.2.4
httplib2 == 0.31.0
idna == 3.10
itsdangerous == 2.2.0
iniconfig == 2.1.0
Jinja2 == 3.1.6
limits == 5.6.0
Mako == 1.3.10
markdown-it-py == 4.0.0
MarkupSafe == 3.0.3
mdurl == 0.1.2
numpy == 2.3.3
oauthlib == 3.3.1
ordered-set == 4.1.0
packaging == 25.0
pluggy == 1.6.0
proto-plus == 1.26.1
protobuf == 6.32.1
pyasn1 == 0.6.1
pyasn1_modules == 0.4.2
pycparser == 2.23
Pygments == 2.19.2
pyparsing == 3.2.5
python-dateutil == 2.9.0.post0
pytz == 2025.2
requests == 2.32.5
requests-oauthlib == 2.0.0
rich == 14.1.0
rsa == 4.9.1
setuptools == 80.9.0
six == 1.17.0
tzdata == 2025.2
typing_extensions == 4.15.0
uritemplate == 4.2.0
urllib3 == 2.5.0
wrapt == 1.17.3
zope.event == 6.0
zope.interface == 8.0.1
